Header refacto to ES6
