+++ src/AppBundle/Resources/public/scripts/main.js
@@ -1,6 +1,12 @@
-import init from './init';
+import init from './global/init';
+import header from './global/header';
 
 // when libraries and template are fully loaded
 $(function() {
+
+    // launch helpers (notifications, slick, reCaptcha, etc.)
     init();
+
+    // header behaviour (search features)
+    header();
 });
